Rename win_api_x86_32_seh -> _structs
This sub-module implements way more than SEH structures
